File Repository Deployment/requirements.txt
Flask==3.0.3
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.12
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
